requirements.txt
streamlit 
langchain 
langchain-openai
validators
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv